requirements.txt
beautifulsoup4>=4.12.3
configparser>=7.1.0
langchain>=0.2.11
langchain-openai>=0.1.6
langchain-core>=0.2.36
langchain-community>=0.2.10
pydantic>=2.7.1
reportlab>=4.2.2
requests>=2.32.3
ruamel.yaml>=0.18.6
pytest>=8.2.2
python-dateutil>=2.8.0
free-proxy>=1.1.1
Jinja2>=3.1.4
PyYAML>=5.3.1
pytest-cov>=5.0.0
PyMuPDF==1.22.5
streamlit